Items can now be added to the playerInventory and moved around in the inventory with the mouse
Added new GameWeapon class to create game weapons. Shop weapon items, when purchased, now use this class to create an instance of the item they represented as an object to be added to the inventory. The inventory is also successfully is saved and loaded with window.localStorage by utilizing func spawnItem() to recreate the items as they were in the inventory before saving / quitting given the saved data